test(devshard): split gateway mockenv coverage
